+++ grumpy/cli/lint.py
@@ -1,0 +1,12 @@
+from cleo import Command
+from ..commands import check_lint
+
+class LintCommand(Command):
+    """
+    Runs lint checks
+
+    lint
+    """
+
+    def handle(self):
+        check_lint()